Move all elements into the same package in order to leverage sealed controller and config interfaces.
