+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/elements/DropdownConfig.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.elements.common
+package com.stripe.android.paymentsheet.elements
 
 internal interface DropdownConfig {
     val debugLabel: String
