+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/elements/DropdownField.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.elements.common
+package com.stripe.android.paymentsheet.elements
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
