+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/elements/SaveForFutureUseController.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.elements.common
+package com.stripe.android.paymentsheet.elements
 
 import com.stripe.android.paymentsheet.R
 import com.stripe.android.paymentsheet.specifications.IdentifierSpec
