+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/elements/TextFieldStateConstants.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.elements.common
+package com.stripe.android.paymentsheet.elements
 
 import androidx.annotation.StringRes
 import com.stripe.android.paymentsheet.R
