+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/elements/SaveForFutureUseControllerTest.kt
@@ -1,6 +1,7 @@
-package com.stripe.android.paymentsheet.elements.common
+package com.stripe.android.paymentsheet.elements
 
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.paymentsheet.elements.SaveForFutureUseController
 import com.stripe.android.paymentsheet.specifications.IdentifierSpec
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
