+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/elements/TextSectionFieldStateConstantsTest.kt
@@ -1,8 +1,8 @@
 package com.stripe.android.paymentsheet.elements
 
 import com.google.common.truth.Truth.assertThat
-import com.stripe.android.paymentsheet.elements.common.TextFieldStateConstants.Error
-import com.stripe.android.paymentsheet.elements.common.TextFieldStateConstants.Valid
+import com.stripe.android.paymentsheet.elements.TextFieldStateConstants.Error
+import com.stripe.android.paymentsheet.elements.TextFieldStateConstants.Valid
 import org.junit.Test
 
 class TextSectionFieldStateConstantsTest {
